+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The major version bump is due to upping the required PHP version from `8.1` to `
 * Added `FileSystem::isExecutable()` method.
 * Added `FileInfo::hasPermissions()` method that accepts an integer or a `Permissions` instance.
 * Added `FileInfo::getPermissions()` method that returns a `Permissions` instance.
+* Added `FileSystem::clearCache()` method.
 
 #### Changes